@@ -3,46 +3,46 @@
     [
       {
         "name": "react-input-autosize",
-        "raw": "react-input-autosize@^1.1.0",
-        "rawSpec": "^1.1.0",
+        "raw": "react-input-autosize@^1.1.3",
+        "rawSpec": "^1.1.3",
         "scope": null,
-        "spec": ">=1.1.0 <2.0.0",
+        "spec": ">=1.1.3 <2.0.0",
         "type": "range"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/react-select"
     ]
   ],
-  "_from": "react-input-autosize@>=1.1.0 <2.0.0",
-  "_id": "react-input-autosize@1.1.0",
+  "_from": "react-input-autosize@>=1.1.3 <2.0.0",
+  "_id": "react-input-autosize@1.1.4",
   "_inCache": true,
   "_installable": true,
   "_location": "/react-input-autosize",
-  "_nodeVersion": "5.0.0",
+  "_nodeVersion": "7.7.1",
   "_npmOperationalInternal": {
     "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/react-input-autosize-1.1.0.tgz_1468157966059_0.25969206308946013"
+    "tmp": "tmp/react-input-autosize-1.1.4.tgz_1494743075192_0.5194365275092423"
   },
   "_npmUser": {
     "email": "jed.watson@me.com",
     "name": "jedwatson"
   },
-  "_npmVersion": "3.3.6",
+  "_npmVersion": "4.4.1",
   "_phantomChildren": {},
   "_requested": {
     "name": "react-input-autosize",
-    "raw": "react-input-autosize@^1.1.0",
-    "rawSpec": "^1.1.0",
+    "raw": "react-input-autosize@^1.1.3",
+    "rawSpec": "^1.1.3",
     "scope": null,
-    "spec": ">=1.1.0 <2.0.0",
+    "spec": ">=1.1.3 <2.0.0",
     "type": "range"
   },
   "_requiredBy": [
     "/react-select"
   ],
-  "_resolved": "https://registry.npmjs.org/react-input-autosize/-/react-input-autosize-1.1.0.tgz",
-  "_shasum": "3fe1ac832387d8abab85f6051ceab1c9e5570853",
+  "_resolved": "https://registry.npmjs.org/react-input-autosize/-/react-input-autosize-1.1.4.tgz",
+  "_shasum": "cbc45072d4084ddc57806db8e3b34e644b8366ac",
   "_shrinkwrap": null,
-  "_spec": "react-input-autosize@^1.1.0",
+  "_spec": "react-input-autosize@^1.1.3",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/react-select",
   "author": {
     "name": "Jed Watson"
@@ -53,13 +53,17 @@
   "bugs": {
     "url": "https://github.com/JedWatson/react-input-autosize/issues"
   },
-  "dependencies": {},
+  "dependencies": {
+    "create-react-class": "^15.5.2",
+    "prop-types": "^15.5.8"
+  },
   "description": "Auto-resizing Input Component for React",
   "devDependencies": {
     "babel-jest": "^6.0.1",
-    "eslint": "^2.11.1",
-    "eslint-config-keystone": "^2.2.0",
-    "eslint-plugin-react": "^5.1.1",
+    "eslint": "^3.19.0",
+    "eslint-config-keystone": "^3.0.0",
+    "eslint-config-keystone-react": "^1.0.0",
+    "eslint-plugin-react": "^6.10.3",
     "gulp": "^3.9.0",
     "jest-cli": "^0.8.1",
     "react": "^15.0",
@@ -69,10 +73,10 @@
   },
   "directories": {},
   "dist": {
-    "shasum": "3fe1ac832387d8abab85f6051ceab1c9e5570853",
-    "tarball": "https://registry.npmjs.org/react-input-autosize/-/react-input-autosize-1.1.0.tgz"
+    "shasum": "cbc45072d4084ddc57806db8e3b34e644b8366ac",
+    "tarball": "https://registry.npmjs.org/react-input-autosize/-/react-input-autosize-1.1.4.tgz"
   },
-  "gitHead": "dbaf3c25b755e1811dc49aec341bec6328cdeef2",
+  "gitHead": "34d4be8c6d9a807a8eb57ddc87d083374b87e00f",
   "homepage": "https://github.com/JedWatson/react-input-autosize#readme",
   "jest": {
     "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
@@ -123,5 +127,5 @@
     "test": "jest; true",
     "watch": "gulp watch:lib"
   },
-  "version": "1.1.0"
+  "version": "1.1.4"
 }
