@@ -1,14 +1,16 @@
 import React from 'react';
+import createClass from 'create-react-class';
+import PropTypes from 'prop-types';
 import Select from 'react-select';
 
 const CONTRIBUTORS = require('../data/contributors');
 const MAX_CONTRIBUTORS = 6;
 const ASYNC_DELAY = 500;
 
-const Contributors = React.createClass({
+const Contributors = createClass({
 	displayName: 'Contributors',
 	propTypes: {
-		label: React.PropTypes.string,
+		label: PropTypes.string,
 	},
 	getInitialState () {
 		return {
