@@ -8,6 +8,10 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _createReactClass = require('create-react-class');
+
+var _createReactClass2 = _interopRequireDefault(_createReactClass);
+
 var _Select = require('./Select');
 
 var _Select2 = _interopRequireDefault(_Select);
@@ -22,8 +26,12 @@ function reduce(obj) {
 	}, props);
 }
 
-var AsyncCreatable = _react2['default'].createClass({
+var AsyncCreatable = (0, _createReactClass2['default'])({
 	displayName: 'AsyncCreatableSelect',
+
+	focus: function focus() {
+		this.select.focus();
+	},
 
 	render: function render() {
 		var _this = this;
@@ -42,6 +50,7 @@ var AsyncCreatable = _react2['default'].createClass({
 								return asyncProps.onInputChange(input);
 							},
 							ref: function (ref) {
+								_this.select = ref;
 								creatableProps.ref(ref);
 								asyncProps.ref(ref);
 							}
