@@ -10,6 +10,14 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _createReactClass = require('create-react-class');
+
+var _createReactClass2 = _interopRequireDefault(_createReactClass);
+
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var _Select = require('./Select');
 
 var _Select2 = _interopRequireDefault(_Select);
@@ -22,52 +30,52 @@ var _utilsDefaultMenuRenderer = require('./utils/defaultMenuRenderer');
 
 var _utilsDefaultMenuRenderer2 = _interopRequireDefault(_utilsDefaultMenuRenderer);
 
-var Creatable = _react2['default'].createClass({
+var Creatable = (0, _createReactClass2['default'])({
 	displayName: 'CreatableSelect',
 
 	propTypes: {
 		// Child function responsible for creating the inner Select component
 		// This component can be used to compose HOCs (eg Creatable and Async)
 		// (props: Object): PropTypes.element
-		children: _react2['default'].PropTypes.func,
+		children: _propTypes2['default'].func,
 
 		// See Select.propTypes.filterOptions
-		filterOptions: _react2['default'].PropTypes.any,
+		filterOptions: _propTypes2['default'].any,
 
 		// Searches for any matching option within the set of options.
 		// This function prevents duplicate options from being created.
 		// ({ option: Object, options: Array, labelKey: string, valueKey: string }): boolean
-		isOptionUnique: _react2['default'].PropTypes.func,
+		isOptionUnique: _propTypes2['default'].func,
 
 		// Determines if the current input text represents a valid option.
 		// ({ label: string }): boolean
-		isValidNewOption: _react2['default'].PropTypes.func,
+		isValidNewOption: _propTypes2['default'].func,
 
 		// See Select.propTypes.menuRenderer
-		menuRenderer: _react2['default'].PropTypes.any,
+		menuRenderer: _propTypes2['default'].any,
 
 		// Factory to create new option.
 		// ({ label: string, labelKey: string, valueKey: string }): Object
-		newOptionCreator: _react2['default'].PropTypes.func,
+		newOptionCreator: _propTypes2['default'].func,
 
 		// input change handler: function (inputValue) {}
-		onInputChange: _react2['default'].PropTypes.func,
+		onInputChange: _propTypes2['default'].func,
 
 		// input keyDown handler: function (event) {}
-		onInputKeyDown: _react2['default'].PropTypes.func,
+		onInputKeyDown: _propTypes2['default'].func,
 
 		// new option click handler: function (option) {}
-		onNewOptionClick: _react2['default'].PropTypes.func,
+		onNewOptionClick: _propTypes2['default'].func,
 
 		// See Select.propTypes.options
-		options: _react2['default'].PropTypes.array,
+		options: _propTypes2['default'].array,
 
 		// Creates prompt/placeholder option text.
 		// (filterText: string): string
-		promptTextCreator: _react2['default'].PropTypes.func,
+		promptTextCreator: _propTypes2['default'].func,
 
 		// Decides if a keyDown event (eg its `keyCode`) should result in the creation of a new option.
-		shouldKeyDownEventCreateNewOption: _react2['default'].PropTypes.func
+		shouldKeyDownEventCreateNewOption: _propTypes2['default'].func
 	},
 
 	// Default prop methods
@@ -221,6 +229,10 @@ var Creatable = _react2['default'].createClass({
 		} else {
 			this.select.selectValue(option);
 		}
+	},
+
+	focus: function focus() {
+		this.select.focus();
 	},
 
 	render: function render() {
