@@ -22,6 +22,14 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _createReactClass = require('create-react-class');
+
+var _createReactClass2 = _interopRequireDefault(_createReactClass);
+
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var _reactDom = require('react-dom');
 
 var _reactDom2 = _interopRequireDefault(_reactDom);
@@ -83,84 +91,85 @@ function stringifyValue(value) {
 	}
 }
 
-var stringOrNode = _react2['default'].PropTypes.oneOfType([_react2['default'].PropTypes.string, _react2['default'].PropTypes.node]);
+var stringOrNode = _propTypes2['default'].oneOfType([_propTypes2['default'].string, _propTypes2['default'].node]);
 
 var instanceId = 1;
 
-var Select = _react2['default'].createClass({
+var Select = (0, _createReactClass2['default'])({
 
 	displayName: 'Select',
 
 	propTypes: {
-		addLabelText: _react2['default'].PropTypes.string, // placeholder displayed when you want to add a label on a multi-value input
-		'aria-label': _react2['default'].PropTypes.string, // Aria label (for assistive tech)
-		'aria-labelledby': _react2['default'].PropTypes.string, // HTML ID of an element that should be used as the label (for assistive tech)
-		arrowRenderer: _react2['default'].PropTypes.func, // Create drop-down caret element
-		autoBlur: _react2['default'].PropTypes.bool, // automatically blur the component when an option is selected
-		autofocus: _react2['default'].PropTypes.bool, // autofocus the component on mount
-		autosize: _react2['default'].PropTypes.bool, // whether to enable autosizing or not
-		backspaceRemoves: _react2['default'].PropTypes.bool, // whether backspace removes an item if there is no text input
-		backspaceToRemoveMessage: _react2['default'].PropTypes.string, // Message to use for screenreaders to press backspace to remove the current item - {label} is replaced with the item label
-		className: _react2['default'].PropTypes.string, // className for the outer element
+		addLabelText: _propTypes2['default'].string, // placeholder displayed when you want to add a label on a multi-value input
+		'aria-describedby': _propTypes2['default'].string, // HTML ID(s) of element(s) that should be used to describe this input (for assistive tech)
+		'aria-label': _propTypes2['default'].string, // Aria label (for assistive tech)
+		'aria-labelledby': _propTypes2['default'].string, // HTML ID of an element that should be used as the label (for assistive tech)
+		arrowRenderer: _propTypes2['default'].func, // Create drop-down caret element
+		autoBlur: _propTypes2['default'].bool, // automatically blur the component when an option is selected
+		autofocus: _propTypes2['default'].bool, // autofocus the component on mount
+		autosize: _propTypes2['default'].bool, // whether to enable autosizing or not
+		backspaceRemoves: _propTypes2['default'].bool, // whether backspace removes an item if there is no text input
+		backspaceToRemoveMessage: _propTypes2['default'].string, // Message to use for screenreaders to press backspace to remove the current item - {label} is replaced with the item label
+		className: _propTypes2['default'].string, // className for the outer element
 		clearAllText: stringOrNode, // title for the "clear" control when multi: true
-		clearRenderer: _react2['default'].PropTypes.func, // create clearable x element
+		clearRenderer: _propTypes2['default'].func, // create clearable x element
 		clearValueText: stringOrNode, // title for the "clear" control
-		clearable: _react2['default'].PropTypes.bool, // should it be possible to reset value
-		deleteRemoves: _react2['default'].PropTypes.bool, // whether backspace removes an item if there is no text input
-		delimiter: _react2['default'].PropTypes.string, // delimiter to use to join multiple values for the hidden field value
-		disabled: _react2['default'].PropTypes.bool, // whether the Select is disabled or not
-		escapeClearsValue: _react2['default'].PropTypes.bool, // whether escape clears the value when the menu is closed
-		filterOption: _react2['default'].PropTypes.func, // method to filter a single option (option, filterString)
-		filterOptions: _react2['default'].PropTypes.any, // boolean to enable default filtering or function to filter the options array ([options], filterString, [values])
-		ignoreAccents: _react2['default'].PropTypes.bool, // whether to strip diacritics when filtering
-		ignoreCase: _react2['default'].PropTypes.bool, // whether to perform case-insensitive filtering
-		inputProps: _react2['default'].PropTypes.object, // custom attributes for the Input
-		inputRenderer: _react2['default'].PropTypes.func, // returns a custom input component
-		instanceId: _react2['default'].PropTypes.string, // set the components instanceId
-		isLoading: _react2['default'].PropTypes.bool, // whether the Select is loading externally or not (such as options being loaded)
-		joinValues: _react2['default'].PropTypes.bool, // joins multiple values into a single form field with the delimiter (legacy mode)
-		labelKey: _react2['default'].PropTypes.string, // path of the label value in option objects
-		matchPos: _react2['default'].PropTypes.string, // (any|start) match the start or entire string when filtering
-		matchProp: _react2['default'].PropTypes.string, // (any|label|value) which option property to filter on
-		menuBuffer: _react2['default'].PropTypes.number, // optional buffer (in px) between the bottom of the viewport and the bottom of the menu
-		menuContainerStyle: _react2['default'].PropTypes.object, // optional style to apply to the menu container
-		menuRenderer: _react2['default'].PropTypes.func, // renders a custom menu with options
-		menuStyle: _react2['default'].PropTypes.object, // optional style to apply to the menu
-		multi: _react2['default'].PropTypes.bool, // multi-value input
-		name: _react2['default'].PropTypes.string, // generates a hidden <input /> tag with this field name for html forms
+		clearable: _propTypes2['default'].bool, // should it be possible to reset value
+		deleteRemoves: _propTypes2['default'].bool, // whether backspace removes an item if there is no text input
+		delimiter: _propTypes2['default'].string, // delimiter to use to join multiple values for the hidden field value
+		disabled: _propTypes2['default'].bool, // whether the Select is disabled or not
+		escapeClearsValue: _propTypes2['default'].bool, // whether escape clears the value when the menu is closed
+		filterOption: _propTypes2['default'].func, // method to filter a single option (option, filterString)
+		filterOptions: _propTypes2['default'].any, // boolean to enable default filtering or function to filter the options array ([options], filterString, [values])
+		ignoreAccents: _propTypes2['default'].bool, // whether to strip diacritics when filtering
+		ignoreCase: _propTypes2['default'].bool, // whether to perform case-insensitive filtering
+		inputProps: _propTypes2['default'].object, // custom attributes for the Input
+		inputRenderer: _propTypes2['default'].func, // returns a custom input component
+		instanceId: _propTypes2['default'].string, // set the components instanceId
+		isLoading: _propTypes2['default'].bool, // whether the Select is loading externally or not (such as options being loaded)
+		joinValues: _propTypes2['default'].bool, // joins multiple values into a single form field with the delimiter (legacy mode)
+		labelKey: _propTypes2['default'].string, // path of the label value in option objects
+		matchPos: _propTypes2['default'].string, // (any|start) match the start or entire string when filtering
+		matchProp: _propTypes2['default'].string, // (any|label|value) which option property to filter on
+		menuBuffer: _propTypes2['default'].number, // optional buffer (in px) between the bottom of the viewport and the bottom of the menu
+		menuContainerStyle: _propTypes2['default'].object, // optional style to apply to the menu container
+		menuRenderer: _propTypes2['default'].func, // renders a custom menu with options
+		menuStyle: _propTypes2['default'].object, // optional style to apply to the menu
+		multi: _propTypes2['default'].bool, // multi-value input
+		name: _propTypes2['default'].string, // generates a hidden <input /> tag with this field name for html forms
 		noResultsText: stringOrNode, // placeholder displayed when there are no matching search results
-		onBlur: _react2['default'].PropTypes.func, // onBlur handler: function (event) {}
-		onBlurResetsInput: _react2['default'].PropTypes.bool, // whether input is cleared on blur
-		onChange: _react2['default'].PropTypes.func, // onChange handler: function (newValue) {}
-		onClose: _react2['default'].PropTypes.func, // fires when the menu is closed
-		onCloseResetsInput: _react2['default'].PropTypes.bool, // whether input is cleared when menu is closed through the arrow
-		onFocus: _react2['default'].PropTypes.func, // onFocus handler: function (event) {}
-		onInputChange: _react2['default'].PropTypes.func, // onInputChange handler: function (inputValue) {}
-		onInputKeyDown: _react2['default'].PropTypes.func, // input keyDown handler: function (event) {}
-		onMenuScrollToBottom: _react2['default'].PropTypes.func, // fires when the menu is scrolled to the bottom; can be used to paginate options
-		onOpen: _react2['default'].PropTypes.func, // fires when the menu is opened
-		onValueClick: _react2['default'].PropTypes.func, // onClick handler for value labels: function (value, event) {}
-		openAfterFocus: _react2['default'].PropTypes.bool, // boolean to enable opening dropdown when focused
-		openOnFocus: _react2['default'].PropTypes.bool, // always open options menu on focus
-		optionClassName: _react2['default'].PropTypes.string, // additional class(es) to apply to the <Option /> elements
-		optionComponent: _react2['default'].PropTypes.func, // option component to render in dropdown
-		optionRenderer: _react2['default'].PropTypes.func, // optionRenderer: function (option) {}
-		options: _react2['default'].PropTypes.array, // array of options
-		pageSize: _react2['default'].PropTypes.number, // number of entries to page when using page up/down keys
+		onBlur: _propTypes2['default'].func, // onBlur handler: function (event) {}
+		onBlurResetsInput: _propTypes2['default'].bool, // whether input is cleared on blur
+		onChange: _propTypes2['default'].func, // onChange handler: function (newValue) {}
+		onClose: _propTypes2['default'].func, // fires when the menu is closed
+		onCloseResetsInput: _propTypes2['default'].bool, // whether input is cleared when menu is closed through the arrow
+		onFocus: _propTypes2['default'].func, // onFocus handler: function (event) {}
+		onInputChange: _propTypes2['default'].func, // onInputChange handler: function (inputValue) {}
+		onInputKeyDown: _propTypes2['default'].func, // input keyDown handler: function (event) {}
+		onMenuScrollToBottom: _propTypes2['default'].func, // fires when the menu is scrolled to the bottom; can be used to paginate options
+		onOpen: _propTypes2['default'].func, // fires when the menu is opened
+		onValueClick: _propTypes2['default'].func, // onClick handler for value labels: function (value, event) {}
+		openAfterFocus: _propTypes2['default'].bool, // boolean to enable opening dropdown when focused
+		openOnFocus: _propTypes2['default'].bool, // always open options menu on focus
+		optionClassName: _propTypes2['default'].string, // additional class(es) to apply to the <Option /> elements
+		optionComponent: _propTypes2['default'].func, // option component to render in dropdown
+		optionRenderer: _propTypes2['default'].func, // optionRenderer: function (option) {}
+		options: _propTypes2['default'].array, // array of options
+		pageSize: _propTypes2['default'].number, // number of entries to page when using page up/down keys
 		placeholder: stringOrNode, // field placeholder, displayed when there's no value
-		required: _react2['default'].PropTypes.bool, // applies HTML5 required attribute when needed
-		resetValue: _react2['default'].PropTypes.any, // value to use when you clear the control
-		scrollMenuIntoView: _react2['default'].PropTypes.bool, // boolean to enable the viewport to shift so that the full menu fully visible when engaged
-		searchable: _react2['default'].PropTypes.bool, // whether to enable searching feature or not
-		simpleValue: _react2['default'].PropTypes.bool, // pass the value to onChange as a simple value (legacy pre 1.0 mode), defaults to false
-		style: _react2['default'].PropTypes.object, // optional style to apply to the control
-		tabIndex: _react2['default'].PropTypes.string, // optional tab index of the control
-		tabSelectsValue: _react2['default'].PropTypes.bool, // whether to treat tabbing out while focused to be value selection
-		value: _react2['default'].PropTypes.any, // initial field value
-		valueComponent: _react2['default'].PropTypes.func, // value component to render
-		valueKey: _react2['default'].PropTypes.string, // path of the label value in option objects
-		valueRenderer: _react2['default'].PropTypes.func, // valueRenderer: function (option) {}
-		wrapperStyle: _react2['default'].PropTypes.object },
+		required: _propTypes2['default'].bool, // applies HTML5 required attribute when needed
+		resetValue: _propTypes2['default'].any, // value to use when you clear the control
+		scrollMenuIntoView: _propTypes2['default'].bool, // boolean to enable the viewport to shift so that the full menu fully visible when engaged
+		searchable: _propTypes2['default'].bool, // whether to enable searching feature or not
+		simpleValue: _propTypes2['default'].bool, // pass the value to onChange as a simple value (legacy pre 1.0 mode), defaults to false
+		style: _propTypes2['default'].object, // optional style to apply to the control
+		tabIndex: _propTypes2['default'].string, // optional tab index of the control
+		tabSelectsValue: _propTypes2['default'].bool, // whether to treat tabbing out while focused to be value selection
+		value: _propTypes2['default'].any, // initial field value
+		valueComponent: _propTypes2['default'].func, // value component to render
+		valueKey: _propTypes2['default'].string, // path of the label value in option objects
+		valueRenderer: _propTypes2['default'].func, // valueRenderer: function (option) {}
+		wrapperStyle: _propTypes2['default'].object },
 
 	// optional style to apply to the component wrapper
 	statics: { Async: _Async2['default'], AsyncCreatable: _AsyncCreatable2['default'], Creatable: _Creatable2['default'] },
@@ -195,7 +204,6 @@ var Select = _react2['default'].createClass({
 			noResultsText: 'No results found',
 			onBlurResetsInput: true,
 			onCloseResetsInput: true,
-			openAfterFocus: false,
 			optionComponent: _Option2['default'],
 			pageSize: 5,
 			placeholder: 'Select...',
@@ -321,12 +329,6 @@ var Select = _react2['default'].createClass({
 	focus: function focus() {
 		if (!this.input) return;
 		this.input.focus();
-
-		if (this.props.openAfterFocus) {
-			this.setState({
-				isOpen: true
-			});
-		}
 	},
 
 	blurInput: function blurInput() {
@@ -940,6 +942,7 @@ var Select = _react2['default'].createClass({
 			'aria-owns': ariaOwns,
 			'aria-haspopup': '' + isOpen,
 			'aria-activedescendant': isOpen ? this._instancePrefix + '-option-' + focusedOptionIndex : this._instancePrefix + '-value',
+			'aria-describedby': this.props['aria-describedby'],
 			'aria-labelledby': this.props['aria-labelledby'],
 			'aria-label': this.props['aria-label'],
 			className: className,
@@ -1111,7 +1114,14 @@ var Select = _react2['default'].createClass({
 
 		var focusedOption = this.state.focusedOption || selectedOption;
 		if (focusedOption && !focusedOption.disabled) {
-			var focusedOptionIndex = options.indexOf(focusedOption);
+			var focusedOptionIndex = -1;
+			options.some(function (option, index) {
+				var isOptionEqual = option.value === focusedOption.value;
+				if (isOptionEqual) {
+					focusedOptionIndex = index;
+				}
+				return isOptionEqual;
+			});
 			if (focusedOptionIndex !== -1) {
 				return focusedOptionIndex;
 			}
@@ -1167,6 +1177,7 @@ var Select = _react2['default'].createClass({
 		var className = (0, _classnames2['default'])('Select', this.props.className, {
 			'Select--multi': this.props.multi,
 			'Select--single': !this.props.multi,
+			'is-clearable': this.props.clearable,
 			'is-disabled': this.props.disabled,
 			'is-focused': this.state.isFocused,
 			'is-loading': this.props.isLoading,
