@@ -1,4 +1,5 @@
 import React from 'react';
+import createClass from 'create-react-class';
 import Select from './Select';
 
 function reduce(obj, props = {}){
@@ -10,8 +11,12 @@ function reduce(obj, props = {}){
   }, props);
 }
 
-const AsyncCreatable = React.createClass({
+const AsyncCreatable = createClass({
 	displayName: 'AsyncCreatableSelect',
+
+	focus () {
+		this.select.focus();
+	},
 
 	render () {
 		return (
@@ -26,6 +31,7 @@ const AsyncCreatable = React.createClass({
 									return asyncProps.onInputChange(input);
 								}}
 								ref={(ref) => {
+									this.select = ref;
 									creatableProps.ref(ref);
 									asyncProps.ref(ref);
 								}}
